# Optimize `excludingPattern` in `violatingOpeningBraceRanges()`