@scottrhoyt 2016-02-06
That's much better. Thanks for fixing this, @norio-nomura !